refactor(gamepad): split ModGamepad.cs into 4 partial class files
ModGamepad.cs        (144 lines) — fields, constants, public API, polling loop
ModGamepadHighlight.cs (178 lines) — overlay creation, highlight positioning, event handlers
ModGamepadInput.cs   (196 lines) — button dispatch, title bar mode, simulate, dialog
ModGamepadNavigation.cs (108 lines) — focus navigation, page tracking, scope stack
