+++ Plain Craft Launcher 2/Modules/ModGamepad.cs
@@ -8,7 +8,7 @@ using SharpDX.XInput;
 
 namespace PCL;
 
-public static class ModGamepad
+public static partial class ModGamepad
 {
     private static Controller _controller;
     private static bool _isRunning;
@@ -92,288 +92,7 @@ public static class ModGamepad
         _pageChildProp.RemoveValueChanged(frm.PanMainRight, OnPageChildChanged);
     }
 
-    private static void CreateHighlightOverlay()
-    {
-        var frm = GetForm();
-        if (frm is null) return;
-
-        _highlightCanvas = new Canvas
-        {
-            IsHitTestVisible = false,
-            Focusable = false
-        };
-        Grid.SetRowSpan(_highlightCanvas, 2);
-
-        _highlightBorderBrush = new SolidColorBrush(Color.FromArgb(200, 0, 120, 215));
-        _highlightBackground = new SolidColorBrush(Color.FromArgb(50, 0, 120, 215));
-        _titleBarBorderBrush = new SolidColorBrush(Color.FromArgb(200, 255, 165, 0));
-        _titleBarBackground = new SolidColorBrush(Color.FromArgb(50, 255, 165, 0));
-
-        _highlightBorder = new Border
-        {
-            IsHitTestVisible = false,
-            BorderBrush = _highlightBorderBrush,
-            BorderThickness = new Thickness(2.5),
-            Background = _highlightBackground,
-            CornerRadius = new CornerRadius(6),
-            Visibility = Visibility.Collapsed,
-            Focusable = false
-        };
-        _highlightCanvas.Children.Add(_highlightBorder);
-
-        var panMainIdx = frm.PanForm.Children.IndexOf(frm.PanMain);
-        frm.PanForm.Children.Insert(panMainIdx + 1, _highlightCanvas);
-
-        _btnTitleMin = frm.FindName("BtnTitleMin") as MyIconButton;
-        _btnTitleHelp = frm.FindName("BtnTitleHelp") as MyIconButton;
-        _btnTitleClose = frm.FindName("BtnTitleClose") as MyIconButton;
-
-        var childProp = DependencyPropertyDescriptor.FromName("Child", typeof(Border), typeof(Border));
-        if (childProp is not null)
-        {
-            _pageChildProp = childProp;
-            childProp.AddValueChanged(frm.PanMainLeft, OnPageChildChanged);
-            childProp.AddValueChanged(frm.PanMainRight, OnPageChildChanged);
-        }
-
-        frm.PreviewMouseDown += OnAnyMouseDown;
-        frm.PreviewMouseMove += OnAnyMouseMove;
-        frm.PreviewKeyDown += OnAnyKeyDown;
-    }
-
-    private static void OnPageChildChanged(object? sender, EventArgs e)
-    {
-        _focusScopeStack.Clear();
-        _pendingPageFocus = true;
-    }
-
-    private static void HideHighlight()
-    {
-        if (_highlightBorder is not null)
-            _highlightBorder.Visibility = Visibility.Collapsed;
-    }
-
     private static FormMain? GetForm() => ModMain.FrmMain;
-
-    private static void OnAnyMouseDown(object sender, MouseButtonEventArgs e)
-    {
-        _gamepadActive = false;
-        HideHighlight();
-    }
-
-    private static void OnAnyMouseMove(object sender, MouseEventArgs e)
-    {
-        var frm = GetForm();
-        if (frm is null) return;
-        var pos = e.GetPosition(frm);
-        var dx = pos.X - _lastMousePos.X;
-        var dy = pos.Y - _lastMousePos.Y;
-        _lastMousePos = pos;
-
-        if (_gamepadActive && (Math.Abs(dx) > MouseHideThreshold || Math.Abs(dy) > MouseHideThreshold))
-        {
-            _gamepadActive = false;
-            HideHighlight();
-        }
-    }
-
-    private static void OnAnyKeyDown(object sender, KeyEventArgs e)
-    {
-        if (_isSimulatingInput) return;
-        _gamepadActive = false;
-        HideHighlight();
-    }
-
-    private static void UpdateHighlightFromFocus()
-    {
-        if (_highlightBorder is null) return;
-
-        if (_isTitleBarMode)
-            UpdateTitleBarHighlight();
-        else if (_gamepadActive)
-            UpdateNormalHighlight();
-        else
-            HideHighlight();
-    }
-
-    private static void UpdateTitleBarHighlight()
-    {
-        var btn = GetTitleBarButton();
-        if (btn is null || !btn.IsVisible || !btn.IsEnabled)
-        {
-            HideHighlight();
-            return;
-        }
-
-        var frm = GetForm();
-        if (frm is null) return;
-
-        _highlightBorder.BorderBrush = _titleBarBorderBrush;
-        _highlightBorder.Background = _titleBarBackground;
-
-        try
-        {
-            var transform = btn.TransformToAncestor(frm.PanForm);
-            var position = transform.Transform(new Point(0, 0));
-
-            _highlightBorder.Width = Math.Max(btn.ActualWidth, 4);
-            _highlightBorder.Height = Math.Max(btn.ActualHeight, 4);
-            _highlightBorder.CornerRadius = new CornerRadius(4);
-
-            Canvas.SetLeft(_highlightBorder, position.X);
-            Canvas.SetTop(_highlightBorder, position.Y);
-            _highlightBorder.Visibility = Visibility.Visible;
-        }
-        catch
-        {
-            HideHighlight();
-        }
-    }
-
-    private static void UpdateNormalHighlight()
-    {
-        if (!_gamepadActive)
-        {
-            HideHighlight();
-            return;
-        }
-
-        var focused = Keyboard.FocusedElement as FrameworkElement;
-        if (focused is null || !focused.IsVisible || !focused.IsEnabled
-            || !KeyControllAbility.GetCanSelect(focused))
-        {
-            HideHighlight();
-            return;
-        }
-
-        var frm = GetForm();
-        if (frm is null) return;
-
-        _highlightBorder.BorderBrush = _highlightBorderBrush;
-        _highlightBorder.Background = _highlightBackground;
-
-        try
-        {
-            var transform = focused.TransformToAncestor(frm.PanForm);
-            var position = transform.Transform(new Point(0, 0));
-
-            _highlightBorder.Width = Math.Max(focused.ActualWidth, 4);
-            _highlightBorder.Height = Math.Max(focused.ActualHeight, 4);
-            _highlightBorder.CornerRadius = focused is Border b
-                ? b.CornerRadius
-                : new CornerRadius(4);
-
-            Canvas.SetLeft(_highlightBorder, position.X);
-            Canvas.SetTop(_highlightBorder, position.Y);
-            _highlightBorder.Visibility = Visibility.Visible;
-        }
-        catch
-        {
-            HideHighlight();
-        }
-    }
-
-    private static MyIconButton? GetTitleBarButton()
-    {
-        var visible = GetVisibleTitleBarButtons();
-        if (visible.Count == 0) return null;
-
-        var idx = visible.IndexOf(_titleBarIndex);
-        return idx >= 0 ? GetButtonByIndex(_titleBarIndex) : GetButtonByIndex(visible[0]);
-    }
-
-    private static void MoveTitleBarIndex(int direction)
-    {
-        var visible = GetVisibleTitleBarButtons();
-        if (visible.Count == 0) return;
-
-        var currentPos = visible.IndexOf(_titleBarIndex);
-        if (currentPos < 0) currentPos = -direction;
-
-        var newPos = (currentPos + direction + visible.Count) % visible.Count;
-        _titleBarIndex = visible[newPos];
-    }
-
-    private static List<int> GetVisibleTitleBarButtons()
-    {
-        var visible = new List<int>(3);
-        if (_btnTitleMin is { Visibility: Visibility.Visible }) visible.Add(0);
-        if (_btnTitleHelp is { Visibility: Visibility.Visible }) visible.Add(1);
-        if (_btnTitleClose is { Visibility: Visibility.Visible }) visible.Add(2);
-        return visible;
-    }
-
-    private static MyIconButton? GetButtonByIndex(int idx) => idx switch
-    {
-        0 => _btnTitleMin,
-        1 => _btnTitleHelp,
-        2 => _btnTitleClose,
-        _ => null
-    };
-
-    private static void EnterTitleBarMode()
-    {
-        _isTitleBarMode = true;
-        _titleBarIndex = 0;
-        _gamepadActive = true;
-        _focusScopeStack.Clear();
-        _preTitleBarFocused = Keyboard.FocusedElement as UIElement;
-
-        var btn = GetTitleBarButton();
-        if (btn is not null)
-            btn.Focus();
-    }
-
-    private static void ExitTitleBarMode()
-    {
-        _isTitleBarMode = false;
-        if (_preTitleBarFocused is { IsVisible: true, IsEnabled: true })
-        {
-            _preTitleBarFocused.Focus();
-            _preTitleBarFocused = null;
-            UpdateHighlightFromFocus();
-            return;
-        }
-        _preTitleBarFocused = null;
-        _pendingPageFocus = true;
-    }
-
-    private static void CheckPageChanged()
-    {
-        if (!_pendingPageFocus)
-            return;
-
-        var frm = GetForm();
-        if (frm is null) return;
-
-        var first = KeyControllAbility.FindFirstSelectable(frm.PageLeft)
-            ?? KeyControllAbility.FindFirstSelectable(frm.PageRight);
-        if (first is not null)
-        {
-            first.Focus();
-            _pendingPageFocus = false;
-            if (_gamepadActive)
-                UpdateHighlightFromFocus();
-        }
-    }
-
-    private static bool IsFocusOnLeftPage()
-    {
-        var frm = GetForm();
-        if (frm is null) return true;
-
-        var focused = Keyboard.FocusedElement as DependencyObject;
-        if (focused is null) return true;
-
-        while (focused is not null)
-        {
-            if (focused == frm.PanMainLeft) return true;
-            if (focused == frm.PanMainRight) return false;
-            focused = VisualTreeHelper.GetParent(focused);
-        }
-
-        return frm.PanMainLeft.Visibility == Visibility.Visible;
-    }
 
     private static void PollLoop()
     {
@@ -461,25 +180,6 @@ public static class ModGamepad
         }
     }
 
-    private static void SimulateScroll(int delta)
-    {
-        if (delta == 0) return;
-        _isSimulatingInput = true;
-
-        var frm = GetForm();
-        if (frm is null) return;
-        var source = PresentationSource.FromVisual(frm);
-        if (source is null) return;
-
-        var target = Keyboard.FocusedElement as UIElement ?? frm;
-        var e = new MouseWheelEventArgs(Mouse.PrimaryDevice, 0, delta)
-        {
-            RoutedEvent = UIElement.MouseWheelEvent
-        };
-        target.RaiseEvent(e);
-        _isSimulatingInput = false;
-    }
-
     private static void RunOnUiThread(Action action)
     {
         GetForm()?.Dispatcher.Invoke(action);
@@ -488,235 +188,4 @@ public static class ModGamepad
     private static bool IsDPadDirection(GamepadButtonFlags btn) => btn
         is GamepadButtonFlags.DPadUp or GamepadButtonFlags.DPadDown
         or GamepadButtonFlags.DPadLeft or GamepadButtonFlags.DPadRight;
-
-    private static void HandleButtonPress(GamepadButtonFlags buttons)
-    {
-        var frm = GetForm();
-        if (frm is null) return;
-        if (!frm.IsActive) return;
-
-        if (_isTitleBarMode)
-        {
-            HandleTitleBarInput(buttons);
-            return;
-        }
-
-        if (frm.PanMsg.Children.Count > 0)
-        {
-            InvokeDialogButton(frm.PanMsg.Children[0], buttons);
-            return;
-        }
-
-        if (IsDPadDirection(buttons))
-        {
-            _gamepadActive = true;
-            _lastMousePos = Mouse.GetPosition(frm);
-
-            if (Keyboard.FocusedElement is ComboBox { IsDropDownOpen: true })
-                return;
-
-            var direction = buttons switch
-            {
-                GamepadButtonFlags.DPadUp => FocusNavigationDirection.Up,
-                GamepadButtonFlags.DPadDown => FocusNavigationDirection.Down,
-                GamepadButtonFlags.DPadLeft => FocusNavigationDirection.Left,
-                _ => FocusNavigationDirection.Right
-            };
-
-            if (!MoveFocus(direction))
-            {
-                if (direction == FocusNavigationDirection.Left)
-                    TryFocusOnPage(frm.PageLeft);
-                else if (direction == FocusNavigationDirection.Right)
-                    TryFocusOnPage(frm.PageRight);
-                else
-                    TryFocusOnPage(IsFocusOnLeftPage() ? frm.PageLeft : frm.PageRight);
-            }
-        }
-        else if (buttons.HasFlag(GamepadButtonFlags.A))
-        {
-            var focused = Keyboard.FocusedElement as FrameworkElement;
-            if (focused is not null && focused.IsVisible && focused.IsEnabled
-                && KeyControllAbility.GetCanActivate(focused))
-            {
-                KeyControllAbility.Activate(focused);
-            }
-            else
-            {
-                _gamepadActive = true;
-                _pendingPageFocus = true;
-            }
-        }
-        else if (buttons.HasFlag(GamepadButtonFlags.B))
-        {
-            if (_focusScopeStack.Count > 0)
-            {
-                var parent = _focusScopeStack.Pop();
-                parent.Focus();
-            }
-            else
-            {
-                SimulateKey(Key.Escape);
-            }
-        }
-        else if (buttons.HasFlag(GamepadButtonFlags.Y))
-        {
-            var focused = Keyboard.FocusedElement as UIElement;
-            if (focused is not null)
-            {
-                var child = KeyControllAbility.FindFirstSelectableChild(focused);
-                if (child is not null)
-                {
-                    _focusScopeStack.Push(focused);
-                    child.Focus();
-                }
-            }
-        }
-        else if (buttons.HasFlag(GamepadButtonFlags.LeftShoulder))
-            NavigateTitleTab(-1);
-        else if (buttons.HasFlag(GamepadButtonFlags.RightShoulder))
-            NavigateTitleTab(1);
-        else if (buttons.HasFlag(GamepadButtonFlags.Start))
-            EnterTitleBarMode();
-        else if (buttons.HasFlag(GamepadButtonFlags.Back))
-        {
-            _gamepadActive = true;
-            _focusScopeStack.Clear();
-            _pendingPageFocus = true;
-
-            var targetPage = IsFocusOnLeftPage()
-                ? (DependencyObject?)frm.PageRight
-                : (DependencyObject?)frm.PageLeft;
-            var first = KeyControllAbility.FindFirstSelectable(targetPage);
-            if (first is not null)
-            {
-                first.Focus();
-                _pendingPageFocus = false;
-            }
-        }
-
-        UpdateHighlightFromFocus();
-    }
-
-    private static void HandleTitleBarInput(GamepadButtonFlags buttons)
-    {
-        if (buttons is GamepadButtonFlags.DPadLeft or GamepadButtonFlags.DPadRight)
-        {
-            MoveTitleBarIndex(buttons == GamepadButtonFlags.DPadRight ? 1 : -1);
-            GetTitleBarButton()?.Focus();
-        }
-        else if (buttons.HasFlag(GamepadButtonFlags.A))
-        {
-            var btn = GetTitleBarButton();
-            if (btn is not null && btn.IsVisible && btn.IsEnabled)
-                KeyControllAbility.Activate(btn);
-            ExitTitleBarMode();
-        }
-        else if (buttons.HasFlag(GamepadButtonFlags.B) || buttons.HasFlag(GamepadButtonFlags.Start))
-        {
-            ExitTitleBarMode();
-        }
-    }
-
-    private static void TryFocusOnPage(DependencyObject? page)
-    {
-        var first = KeyControllAbility.FindFirstSelectable(page);
-        if (first is null)
-        {
-            var frm = GetForm();
-            var opposite = ReferenceEquals(page, frm?.PageLeft)
-                ? (DependencyObject?)frm?.PageRight
-                : (DependencyObject?)frm?.PageLeft;
-            if (opposite != page)
-                first = KeyControllAbility.FindFirstSelectable(opposite);
-        }
-        first?.Focus();
-    }
-
-    private static bool MoveFocus(FocusNavigationDirection direction)
-    {
-        var request = new TraversalRequest(direction);
-        var prevFocus = Keyboard.FocusedElement;
-        var focused = prevFocus as UIElement;
-
-        bool moved;
-        if (focused is not null)
-        {
-            moved = focused.MoveFocus(request);
-        }
-        else
-        {
-            var frm = GetForm();
-            moved = frm is not null && ((UIElement)frm).MoveFocus(request);
-        }
-
-        if (!moved)
-            return false;
-
-        var newFocus = Keyboard.FocusedElement;
-        return newFocus != prevFocus
-            && newFocus is FrameworkElement fe
-            && KeyControllAbility.GetCanSelect(fe);
-    }
-
-    private static void NavigateTitleTab(int direction)
-    {
-        var frm = GetForm();
-        if (frm is null) return;
-
-        if (frm.PanTitleMain.Visibility != Visibility.Visible)
-            return;
-
-        var buttons = frm.PanTitleSelect.Children;
-        var currentIndex = 0;
-        for (var i = 0; i < buttons.Count; i++)
-        {
-            if (buttons[i] is MyRadioButton { Checked: true })
-            {
-                currentIndex = i;
-                break;
-            }
-        }
-
-        var newIndex = (currentIndex + direction + buttons.Count) % buttons.Count;
-        if (buttons[newIndex] is MyRadioButton target)
-            target.SetChecked(true, true, true);
-    }
-
-    private static void SimulateKey(Key key)
-    {
-        _isSimulatingInput = true;
-        var frm = GetForm();
-        if (frm is null) return;
-        var source = PresentationSource.FromVisual(frm);
-        if (source is null) return;
-
-        var e = new KeyEventArgs(Keyboard.PrimaryDevice, source, 0, key)
-        {
-            RoutedEvent = Keyboard.KeyDownEvent
-        };
-        frm.RaiseEvent(e);
-        _isSimulatingInput = false;
-    }
-
-    private static void InvokeDialogButton(UIElement msg, GamepadButtonFlags buttons)
-    {
-        Action? action = (buttons, msg) switch
-        {
-            (GamepadButtonFlags.A, MyMsgInput input) => () => input.Btn1_Click(null, null),
-            (GamepadButtonFlags.A, MyMsgSelect select) => () => select.Btn1_Click(null, null),
-            (GamepadButtonFlags.A, MyMsgText text) => () => text.Btn1_Click(),
-            (GamepadButtonFlags.A, MyMsgMarkdown markdown) => () => markdown.Btn1_Click(null, null),
-            (GamepadButtonFlags.A, MyMsgLogin login) => () => login.Btn1_Click(null, null),
-            (GamepadButtonFlags.B, MyMsgInput input) => () => input.Btn2_Click(null, null),
-            (GamepadButtonFlags.B, MyMsgSelect select) => () => select.Btn2_Click(null, null),
-            (GamepadButtonFlags.B, MyMsgText text) => () => text.Btn2_Click(null, null),
-            (GamepadButtonFlags.B, MyMsgMarkdown markdown) => () => markdown.Btn2_Click(null, null),
-            (GamepadButtonFlags.X, MyMsgText text) => () => text.Btn3_Click(null, null),
-            (GamepadButtonFlags.X, MyMsgMarkdown markdown) => () => markdown.Btn3_Click(null, null),
-            (GamepadButtonFlags.X, MyMsgLogin login) => () => login.Btn3_Click(null, null),
-            _ => null
-        };
-        action?.Invoke();
-    }
 }
